+++ src/test/java/org/mastodon/mamut/feature/branch/successors/BranchNSuccessorsFeatureComputerTest.java
@@ -2,7 +2,7 @@ package org.mastodon.mamut.feature.branch.successors;
 
 import org.junit.Test;
 import org.mastodon.feature.FeatureProjection;
-import org.mastodon.mamut.feature.branch.FeatureComputerTestUtils;
+import org.mastodon.mamut.feature.FeatureComputerTestUtils;
 import org.mastodon.mamut.feature.branch.exampleGraph.ExampleGraph1;
 import org.mastodon.mamut.feature.branch.exampleGraph.ExampleGraph2;
 import org.mastodon.mamut.model.branch.BranchSpot;
@@ -19,7 +19,7 @@ public class BranchNSuccessorsFeatureComputerTest
 		{
 			ExampleGraph1 exampleGraph1 = new ExampleGraph1();
 			FeatureProjection< BranchSpot > featureProjection =
-					FeatureComputerTestUtils.getBranchSpotFeatureProjection( context, exampleGraph1,
+					FeatureComputerTestUtils.getFeatureProjection( context, exampleGraph1,
 							BranchNSuccessorsFeature.BRANCH_N_SUCCESSORS_FEATURE,
 							BranchNSuccessorsFeature.PROJECTION_SPEC );
 			assertEquals( 0, featureProjection.value( exampleGraph1.branchSpotA ), 0 );
@@ -33,7 +33,7 @@ public class BranchNSuccessorsFeatureComputerTest
 		{
 			ExampleGraph2 exampleGraph2 = new ExampleGraph2();
 			FeatureProjection< BranchSpot > featureProjection =
-					FeatureComputerTestUtils.getBranchSpotFeatureProjection( context, exampleGraph2,
+					FeatureComputerTestUtils.getFeatureProjection( context, exampleGraph2,
 							BranchNSuccessorsFeature.BRANCH_N_SUCCESSORS_FEATURE,
 							BranchNSuccessorsFeature.PROJECTION_SPEC );
